+++ sdk-clients/tokens/tokens_types_extended.go
@@ -15,18 +15,6 @@ type ProviderTokenDtoFixed struct {
 	Tags            []TagDto `json:"tags"`
 }
 
-// TokenInfoDtoFixed uses the Tag struct for Tags
-type TokenInfoDtoFixed struct {
-	Address    string                  `json:"address"`
-	ChainId    float32                 `json:"chainId"`
-	Decimals   float32                 `json:"decimals"`
-	Extensions *map[string]any `json:"extensions,omitempty"`
-	LogoURI    string                  `json:"logoURI"`
-	Name       string                  `json:"name"`
-	Symbol     string                  `json:"symbol"`
-	Tags       []TagDto                `json:"tags"`
-}
-
 type CustomTokensControllerGetTokenInfoParams struct {
 	Address string `url:"address" json:"address"`
 }